+++ Assets/Scripts/Controller/PlayerController.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using UnityEngine;
+
+
+public class PlayerController : Player
+{
+
+}
