+++ Assets/Scripts/Model/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53902a546ec309e419b5a3ee81f65877
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
