+++ Assets/Scripts/MyPlayer.cs
@@ -4,21 +4,21 @@ using UnityEngine;
 using System;
 using UnityEngine.SceneManagement;
 
-public class MyPlayer : MonoBehaviour, iTakeDamage, iHealth
+public class MyPlayer : MonoBehaviour
 {
-    public static MyPlayer myPlayer;
-    public SlideHealth healthBar;
+    //public static MyPlayer myPlayer;
+    //public SlideHealth healthBar;
 
-    #region Properties
-    [SerializeField] public int currentHp;
-    [SerializeField] public int maxHp;
-    [SerializeField] public float speed = 9f;
-    [SerializeField] private float jumpForce = 300f;
-    [SerializeField] private float rotationSpeed = 0.8f;
-    [SerializeField] private int damage = 10;
+    //#region Properties
+    //[SerializeField] public int currentHp;
+    //[SerializeField] public int maxHp;
+    //[SerializeField] public float speed = 9f;
+    //[SerializeField] private float jumpForce = 300f;
+    //[SerializeField] private float rotationSpeed = 0.8f;
+    //[SerializeField] private int damage = 10;
 
-    Vector3 direction = Vector3.zero;
-    private AudioSource audioSource = null;
+    //Vector3 direction = Vector3.zero;
+    //private AudioSource audioSource = null;
 
     [Header("Bullet")]
     [SerializeField] private float reloadTime = 1.0f;
@@ -33,38 +33,38 @@ public class MyPlayer : MonoBehaviour, iTakeDamage, iHealth
     private float currentReloadTime = 0;
 
 
-    #endregion
+    //#endregion
 
-    private void Awake()
-    {
-        maxHp = 100;
-        myPlayer = this;
-        audioSource = GetComponent<AudioSource>();
-    }
+    //private void Awake()
+    //{
+    //    maxHp = 100;
+    //    myPlayer = this;
+    //    audioSource = GetComponent<AudioSource>();
+    //}
 
-    private void Start()
-    {
-        currentHp = maxHp;
-        healthBar.SetMaxHealth(maxHp);
-    }
+    //private void Start()
+    //{
+    //    currentHp = maxHp;
+    //    healthBar.SetMaxHealth(maxHp);
+    //}
 
-    private void Update()
-    {
-        MovementLogic();
-        Shot();
-    }
+    //private void Update()
+    //{
+    //    MovementLogic();
+    //    Shot();
+    //}
 
-    private void MovementLogic()
-    {
-        direction.z = Input.GetAxis("Vertical") * speed;
-        direction.x = Input.GetAxis("Horizontal") * rotationSpeed;
-        direction.y = Input.GetAxis("Jump");
-        var rotation = direction.x * Time.deltaTime;
-        var move = direction.z * Time.deltaTime;
-        var jump = direction.y;
-        transform.Translate(0, 0, move);
-        transform.Rotate(0, rotation, 0);
-    }
+    //private void MovementLogic()
+    //{
+    //    direction.z = Input.GetAxis("Vertical") * speed;
+    //    direction.x = Input.GetAxis("Horizontal") * rotationSpeed;
+    //    direction.y = Input.GetAxis("Jump");
+    //    var rotation = direction.x * Time.deltaTime;
+    //    var move = direction.z * Time.deltaTime;
+    //    var jump = direction.y;
+    //    transform.Translate(0, 0, move);
+    //    transform.Rotate(0, rotation, 0);
+    //}
 
        
     private void Shot()
@@ -75,7 +75,7 @@ public class MyPlayer : MonoBehaviour, iTakeDamage, iHealth
             {
                 var createBullet = Instantiate(bullet, bulletStartPosition.position, transform.rotation).GetComponent<BulletPlayer>();
                 createBullet.Init();
-                audioSource.Play();
+                //audioSource.Play();
                 currentReloadTime = reloadTime;
             }
         }
@@ -88,28 +88,28 @@ public class MyPlayer : MonoBehaviour, iTakeDamage, iHealth
 
         if (currentReloadTime > 0) currentReloadTime -= Time.deltaTime;
     }
-    public void TakeDamage(int dam)
-    {
-        currentHp -= dam;
-        healthBar.SetHealth(currentHp);
-        if (currentHp <= 0) Death();
-    }
+    //public void TakeDamage(int dam)
+    //{
+    //    currentHp -= dam;
+    //    healthBar.SetHealth(currentHp);
+    //    if (currentHp <= 0) Death();
+    //}
 
-    public void Death()
-    {
-        Destroy(gameObject);
-    }
+    //public void Death()
+    //{
+    //    Destroy(gameObject);
+    //}
 
-    public void Health(int hp)
-    {
-        if (hp + currentHp >= maxHp)
-        {
-            currentHp = maxHp;
-        }
-        else
-        {
-            currentHp += hp;
-        }
-    }
+    //public void Health(int hp)
+    //{
+    //    if (hp + currentHp >= maxHp)
+    //    {
+    //        currentHp = maxHp;
+    //    }
+    //    else
+    //    {
+    //        currentHp += hp;
+    //    }
+    //}
 }
