--- Assets/Scripts/TestMatch.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cda0233cd75f9124bb49516b202811fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
