--- Assets/Scripts/View/BulletView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c5e7fbaf0480df43bbe43a562d73e35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
