+++ Library/EditorOnlyScriptingSettings.json
@@ -1,1 +1,1 @@
-{"m_DefineSymbols":{"m_Value":[],"m_Initialized":true},"m_AllowUnsafeCode":{"m_Value":false,"m_Initialized":true},"m_ScriptDebugInfoEnabled":{"m_Value":false,"m_Initialized":true}}
+{"m_DefineSymbols":{"m_Value":[],"m_Initialized":true},"m_AllowUnsafeCode":{"m_Value":false,"m_Initialized":true},"m_ScriptDebugInfoEnabled":{"m_Value":true,"m_Initialized":true}}
