+++ Logs/ApiUpdaterCheck.txt
@@ -78,3 +78,43 @@ C# parse time         : 249ms
 candidates check time : 67ms
 console write time    : 0ms
 
+[api-updater (non-obsolete-error-filter)] 12.04.2021 16:25:12 : Starting E:/Revina/2020.3.1f1/Editor/Data/Tools/ScriptUpdater/APIUpdater.NonObsoleteApiUpdaterDetector.exe
+[api-updater (non-obsolete-error-filter)] 
+----------------------------------
+jit/startup time      : 1516,7805ms
+moved types parse time: 133ms
+candidates parse time : 1ms
+C# parse time         : 1229ms
+candidates check time : 63ms
+console write time    : 6ms
+
+[api-updater (non-obsolete-error-filter)] 12.04.2021 17:10:52 : Starting E:/Revina/2020.3.1f1/Editor/Data/Tools/ScriptUpdater/APIUpdater.NonObsoleteApiUpdaterDetector.exe
+[api-updater (non-obsolete-error-filter)] 
+----------------------------------
+jit/startup time      : 396,7549ms
+moved types parse time: 64ms
+candidates parse time : 2ms
+C# parse time         : 231ms
+candidates check time : 61ms
+console write time    : 0ms
+
+[api-updater (non-obsolete-error-filter)] 12.04.2021 17:37:08 : Starting E:/Revina/2020.3.1f1/Editor/Data/Tools/ScriptUpdater/APIUpdater.NonObsoleteApiUpdaterDetector.exe
+[api-updater (non-obsolete-error-filter)] 
+----------------------------------
+jit/startup time      : 428,7365ms
+moved types parse time: 81ms
+candidates parse time : 1ms
+C# parse time         : 438ms
+candidates check time : 55ms
+console write time    : 0ms
+
+[api-updater (non-obsolete-error-filter)] 12.04.2021 17:37:28 : Starting E:/Revina/2020.3.1f1/Editor/Data/Tools/ScriptUpdater/APIUpdater.NonObsoleteApiUpdaterDetector.exe
+[api-updater (non-obsolete-error-filter)] 
+----------------------------------
+jit/startup time      : 94,941ms
+moved types parse time: 77ms
+candidates parse time : 1ms
+C# parse time         : 250ms
+candidates check time : 60ms
+console write time    : 0ms
+
